Generalized data loading.
Began generalization of model loading. The goals it to be able to load a big or small model based on the same parameters.
With these changes I have separated out the construction of the big model from the parameters. Next is to make something that constructs a small model from the same set of parameters.